Gulp Rules
Managed to automate Gulp tasks
Tasks.json has KarmaTask, ExplicitKarmaTask (different Gulp setup but
does the same), Protractor and httpServer.
Karma tasks are for Karma tests, positioned in Root/test/unit/**Spec.js
httpServer task starts the http-server. It is supposed to be used
instead of cmd -> server.bat  which starts the application in express
server at localhost:8000
Protractor is for protractor tests. When Protractor task is started,
first the httpServer starts and then Protractor. That is because
Protractor needs the httpServer running.